llm_service/requirements.txt
grpcio-tools~=1.67.0
aiohttp~=3.10.10
aiohttp_socks~=0.9.0
grpcio~=1.67.0
loguru~=0.7.2
pydantic-settings~=2.5.2
